[x86] spinlock: make ticket lock increment 2, unconditionally
Message-id: <20130725114957.5aa230ac@cuia.bos.redhat.com>
Patchwork-id: 64536
O-Subject: [RHEL7 PATCH 2/1] make ticket lock increment 2, unconditionally
Bugzilla: 970737
RH-Acked-by: Andrew Jones <drjones@redhat.com>

Because CONFIG_PARAVIRT_SPINLOCKS is not set, we cannot depend on
that to get an increment of 2 for the ticket locks.

This patch makes the ticket increment 2 unconditionally.

We will have to find a clever way to integrate a paravirt spinlock
solution later...

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>