[powerpc] perf: Check that events only include valid bits on Power8
Message-id: <20130729000219.47366.93955.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 64611
O-Subject: [PATCH RHEL7.0 BZ969176 1/7] powerpc/perf: Check that events only include valid bits on Power8
Bugzilla: 969176
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=969176

Description:
------------
A mistake we have made in the past is that we pull out the fields we
need from the event code, but don't check that there are no unknown bits
set. This means that we can't ever assign meaning to those unknown bits
in future.

Although we have once again failed to do this at release, it is still
early days for Power8 so I think we can still slip this in and get away
with it.

Signed-off-by: Michael Ellerman <michael@ellerman.id.au>
CC: <stable@vger.kernel.org> [v3.10]
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=d8bec4c9cd58f6d3679e09b7293851fb92ad7557

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>