[redhat] Kconfig: Stop compiling the NFS v2 kernel module
Message-id: <51F6541E.9060809@RedHat.com>
Patchwork-id: 64626
O-Subject: [RHEL7.0 PATCH] Stop compiling the NFS v2 kernel module
Bugzilla: 989238
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

[Resending adding the NFS Team alias ]

The following patch disables the NFS v2 kernel module
from being compiled with which in turn disable
the NFS v2 support in RHEL7.

NFS v2 is no longer used and should not be used by
any type of enterprise environment. So the testing
cycles and debugging bug found by that testing is
effort is a complete was of resources.

Signed-off-by: Steve Dickson <steved@redhat.com>
Red Hat BZ: 989238

Signed-off-by: Jarod Wilson <jarod@redhat.com>