[fs] nfsd: allow destroy_session over destroyed session
Message-id: <1375098022-9479-31-git-send-email-steved@redhat.com>
Patchwork-id: 64689
O-Subject: [RHEL7.0 PATCH 30/36] nfsd4: allow destroy_session over destroyed session
Bugzilla: 987090
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: "J. Bruce Fields" <bfields@redhat.com>

RFC 5661 allows a client to destroy a session using a compound
associated with the destroyed session, as long as the DESTROY_SESSION op
is the last op of the compound.

We attempt to allow this, but testing against a Solaris client (which
does destroy sessions in this way) showed that we were failing the
DESTROY_SESSION with NFS4ERR_DELAY, because we assumed the reference
count on the session (held by us) represented another rpc in progress
over this session.

Fix this by noting that in this case the expected reference count is 1,
not 0.

Also, note as long as the session holds a reference to the compound
we're destroying, we can't free it here--instead, delay the free till
the final put in nfs4svc_encode_compoundres.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream ID: f0f51f5cdd107971282ae18f00a6fa03d69407a0
Red Hat BZ: 987090
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>