[powerpc] tm: Fix writing top half of MSR on 32 bit signals
Message-id: <20130729145330.7030.67179.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 64752
O-Subject: [PATCH RHEL7.0 BZ731886 1/5] powerpc/tm: Fix writing top half of MSR on 32 bit signals
Bugzilla: 731886
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=731886

Description:
------------
The MSR TM controls are in the top 32 bits of the MSR hence on 32 bit signals,
we stick the top half of the MSR in the checkpointed signal context so that the
user can access it.

Unfortunately, we don't currently write anything to the checkpointed signal
context when coming in a from a non transactional process and hence the top MSR
bits can contain junk.

This updates the 32 bit signal handling code to always write something to the
top MSR bits so that users know if the process is transactional or not and the
kernel can use it on signal return.

Signed-off-by: Michael Neuling <mikey@neuling.org>
cc: stable@vger.kernel.org (v3.9+)
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=1d25f11fdbcc5390d68efd98c28900bfd29b264c

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>