[powerpc] tm: Fix restoration of MSR on 32bit signal return
Message-id: <20130729145342.7030.67781.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 64754
O-Subject: [PATCH RHEL7.0 BZ731886 3/5] powerpc/tm: Fix restoration of MSR on 32bit signal return
Bugzilla: 731886
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=731886

Description:
------------
Currently we clear out the MSR TM bits on signal return assuming that the
signal should never return to an active transaction.

This is bogus as the user may do this.  It's most likely the transaction will
be doomed due to a treclaim but that's a problem for the HW not the kernel.

The current code is a legacy of earlier kernel implementations which did
software rollback of active transactions in the kernel.  That code has now gone
but we didn't correctly fix up this part of the signals code which still makes
the assumption that it must be returning to a suspended transaction.

This pulls out both MSR TM bits from the user supplied context rather than just
setting TM suspend.  We pull out only the bits needed to ensure the user can't
do anything dangerous to the MSR.

Signed-off-by: Michael Neuling <mikey@neuling.org>
cc: stable@vger.kernel.org (v3.9+)
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=2c27a18f8736da047bef2b997bdd48efc667e3c9

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>