[fs] gfs2: Only do one directory search on create
Message-id: <1375183397.2725.17.camel@menhir>
Patchwork-id: 64797
O-Subject: [RHEL7] GFS2: Add atomic_open feature (bz #983098) [1/2]
Bugzilla: 983098
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is the first of a couple of patches which add the atomic_open
feature to GFS2. The upstream patch is here:

http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2/inode.c?id=5a00f3cc978be45b9d2597851bedaa40630bc597

The only difference between this patch and upstream is that PTR_RET was
written inline rather than using the PTR_RET macro in this case, so the
context was slightly different, even though functionally the same.

This is for bug #983098

Signed-off-by: Jarod Wilson <jarod@redhat.com>