[powerpc] mm/numa: VPHN topology change updates all siblings
Message-id: <20130802122811.8286.90037.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 65038
O-Subject: [PATCH RHEL7.0 BZ973594] powerpc: VPHN topology change updates all siblings
Bugzilla: 973594
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=979523

Description:
------------
When an associativity level change is found for one thread, the
siblings threads need to be updated as well.  This is done today
for PRRN in stage_topology_update() but is missing for VPHN in
update_cpu_associativity_changes_mask().  This patch will correctly
update all thread siblings during a topology change.

Without this patch a topology update can result in a CPU in
init_sched_groups_power() getting stuck indefinitely in a loop.

This loop is built in build_sched_groups(). As a result of the thread
moving to a node separate from its siblings the struct sched_group will
have its next pointer set to point to itself rather than the sched_group
struct of the next thread.  This happens because we have a domain without
the SD_OVERLAP flag, which is correct, and a topology that doesn't conform
with reality (threads on the same core assigned to different numa nodes).
When this list is traversed by init_sched_groups_power() it will reach
the thread's sched_group structure and loop indefinitely; the cpu will
be stuck at this point.

The bug was exposed when VPHN was enabled in commit b7abef0 (v3.9).

Cc: <stable@vger.kernel.org> [v3.9+]
Reported-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Robert Jennings <rcj@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 7.0

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6118564

Upstream:
---------
http://git.kernel.org/cgit/linux/kernel/git/benh/powerpc.git/commit/?h=merge&id=3be7db6ab45b21345386d1a466da133b19cde5e4

Test Status:
------------
Jan Stancek and Robert Jennings helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>