[md] raid0: fix buglet in RAID5 -> RAID0 conversion
Message-id: <1375952433-16555-7-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 65165
O-Subject: [RHEL7.0 PATCH 06/15] md: fix buglet in RAID5 -> RAID0 conversion.
Bugzilla: 978055
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

RAID5 uses a 'per-array' value for the 'size' of each device.
RAID0 uses a 'per-device' value - it can be different for each device.

When converting a RAID5 to a RAID0 we must ensure that the per-device
size of each device matches the per-array size for the RAID5, else
the array will change size.

If the metadata cannot record a changed per-device size (as is the
case with v0.90 metadata) the array could get bigger on restart.  This
does not cause data corruption, so it not a big issue and is mainly
yet another a reason to not use 0.90.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit eea136d69f9facb2d3807386bbac1e6b1161795f)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>