[md] Remember the last sync operation that was performed
Message-id: <1375952433-16555-8-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 65167
O-Subject: [RHEL7.0 PATCH 07/15] MD: Remember the last sync operation that was performed
Bugzilla: 978055
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Jonathan Brassow <jbrassow@redhat.com>

MD:  Remember the last sync operation that was performed

This patch adds a field to the mddev structure to track the last
sync operation that was performed.  This is especially useful when
it comes to what is recorded in mismatch_cnt in sysfs.  If the
last operation was "data-check", then it reports the number of
descrepancies found by the user-initiated check.  If it was a
"repair" operation, then it is reporting the number of
descrepancies repaired.  etc.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit c4a39551451666229b4ea5e8aae8ca0131d00665)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>