[md] raid5: allow 5-device RAID6 to be reshaped to 4-device
Message-id: <1375952433-16555-10-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 65170
O-Subject: [RHEL7.0 PATCH 09/15] md/raid5: allow 5-device RAID6 to be reshaped to 4-device.
Bugzilla: 978055
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

There is a bug in 'check_reshape' for raid5.c  To checks
that the new minimum number of devices is large enough (which is
good), but it does so also after the reshape has started (bad).

This is bad because
 - the calculation is now wrong as mddev->raid_disks has changed
   already, and
 - it is pointless because it is now too late to stop.

So only perform that test when reshape has not been committed to.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit fdcfbbb653b27964c4daa4d2bcb364259c257e7d)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>