[md] raid1: fix bio handling problems in process_checks()
Message-id: <1375952433-16555-14-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 65175
O-Subject: [RHEL7.0 PATCH 13/15] md/raid1: fix bio handling problems in process_checks()
Bugzilla: 978055
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

Recent change to use bio_copy_data() in raid1 when repairing
an array is faulty.

The underlying may have changed the bio in various ways using
bio_advance and these need to be undone not just for the 'sbio' which
is being copied to, but also the 'pbio' (primary) which is being
copied from.

So perform the reset on all bios that were read from and do it early.

This also ensure that the sbio->bi_io_vec[j].bv_len passed to
memcmp is correct.

This fixes a crash during a 'check' of a RAID1 array.  The crash was
introduced in 3.10 so this is suitable for 3.10-stable.

Cc: stable@vger.kernel.org (3.10)
Reported-by: Joe Lawrence <joe.lawrence@stratus.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 30bc9b53878a9921b02e3b5bc4283ac1c6de102a)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>