[base] power/wakeup: Adjust messaging for wake events during suspend
Message-id: <20130808170810.20458.74756.stgit@amt.stowe>
Patchwork-id: 65274
O-Subject: [RHEL7 PATCH 077/153] PM / wakeup: Adjust messaging for wake events during suspend
Bugzilla: 991615
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 991615
https://bugzilla.redhat.com/show_bug.cgi?id=991615
Upstream Status: 9350de06be45a5a8b927ac6577c9d35de61c90ca

commit 9350de06be45a5a8b927ac6577c9d35de61c90ca
Author: Bernie Thompson <bhthompson@chromium.org>
Date:   Sat Jun 1 00:47:43 2013 +0000

PM / wakeup: Adjust messaging for wake events during suspend

This adds in a new message to the wakeup code which adds an
indication to the log that suspend was cancelled due to a wake event
occouring during the suspend sequence. It also adjusts the message
printed in suspend.c to reflect the potential that a suspend was
aborted, as opposed to a device failing to suspend.

Without these message adjustments one can end up with a kernel log
that says that a device failed to suspend with no actual device
suspend failures, which can be confusing to the log examiner.

Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>