[redhat] kernel.spec: hackery to process configs
This whole mess is a house of cards. Change one thing, and seventeen
others cascadingly fail. Adding the symlink to git made the build fail, so
as a quick hack-around, we'll have the rpm %prep remove the symlink and
make a directory in its place where it can stash processed configs.

Signed-off-by: Jarod Wilson <jarod@redhat.com>