[net] openvswitch: Immediately exit on error in ovs_vport_cmd_set()
Message-id: <1376052430-15968-5-git-send-email-amwang@redhat.com>
Patchwork-id: 65364
O-Subject: [RHEL7 Patch net 04/38] openvswitch: Immediately exit on error in ovs_vport_cmd_set().
Bugzilla: 992917
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

It is an error to try to change the type of a vport using the set
command. However, while we check that this is an error, we still
proceed to allocate memory which then gets freed immediately.
This stops processing after noticing the error, which does not
actually fix a bug but is more correct.

Signed-off-by: Jesse Gross <jesse@nicira.com>
(cherry picked from commit f44f340883388b57fe03edfb0982e038e57a992c)

Signed-off-by: Cong Wang <amwang@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>