[net] sit: add support of x-netns
Message-id: <1376052430-15968-30-git-send-email-amwang@redhat.com>
Patchwork-id: 65388
O-Subject: [RHEL7 Patch net 29/38] sit: add support of x-netns
Bugzilla: 992917
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

This patch allows to switch the netns when packet is encapsulated or
decapsulated. In other word, the encapsulated packet is received in a netns,
where the lookup is done to find the tunnel. Once the tunnel is found, the
packet is decapsulated and injecting into the corresponding interface which
stands to another netns.

When one of the two netns is removed, the tunnel is destroyed.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 5e6700b3bf98fe98d630bf9c939ad4c85ce95592)

Signed-off-by: Cong Wang <amwang@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>