[fs] gfs2: don't overrun reserved revokes
Message-id: <20130813042825.GZ2113@dhcp80-209.msp.redhat.com>
Patchwork-id: 65445
O-Subject: [RHEL 7.0 PATCH] GFS2: don't overrun reserved revokes
Bugzilla: 950622
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When run during fsync, a gfs2_log_flush could happen between the
time when gfs2_ail_flush checked the number of blocks to revoke,
and when it actually started the transaction to do those revokes.
This occassionally caused it to need more revokes than it reserved,
causing gfs2 to crash.

Instead of just reserving enough revokes to handle the blocks that
currently need them, this patch makes gfs2_ail_flush reserve the
maximum number of revokes it can, without increasing the total number
of reserved log blocks. This patch also passes the number of reserved
revokes to __gfs2_ail_flush() so that it doesn't go over its limit
and cause a crash like we're seeing. Non-fsync calls to __gfs2_ail_flush
will still cause a BUG() necessary revokes are skipped. The order that
the revokes are reserved has also been switched to do the oldest revokes
first.  This guarantees that the only blocks which might not get revoked
in a fsync were written out after the start of the gfs2_ail_flush().

This patch is the fix for BZ #950622

This fix has also been submitted to Steve's upstream git tree

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>