[powerpc] pseries: Inform the hypervisor we are using EBB regs
Message-id: <20130820190223.40357.57258.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 65631
O-Subject: [PATCH RHEL7.0 BZ997646 3/3] powerpc/pseries: Inform the hypervisor we are using EBB regs
Bugzilla: 997646
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=997646

Description:
------------
On LPAR systems we need to inform the hypervisor that we are using the
EBB registers. We do this by setting a bit in the Virtual Processor Area
(VPA) - formerly known as the lppaca.

For now we do this always, ie. we do not dynamically enable/disable.

Signed-off-by: Michael Ellerman <michael@ellerman.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=6e0b8bc965d25a8e0701eaca3fca5941b4f4b2b2

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>