[powerpc] Avoid link stack corruption for MMU on exceptions
Message-id: <20130822093757.11975.5891.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 65805
O-Subject: [PATCH RHEL7.0 BZ999556] powerpc: Avoid link stack corruption for MMU on exceptions
Bugzilla: 999556
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=999556

Description:
------------
When we have MMU on exceptions (POWER8) and a relocatable kernel, we
need to branch from the initial exception vectors at 0x0 to up high
where the kernel might be located.  Currently we do this using the link
register.

Unfortunately this corrupts the link stack and instead we should use the
count register.  We did this for the syscall entry path in:
  6a40480 powerpc: Avoid link stack corruption in MMU on syscall entry path
but I stupidly forgot to do the same for other exceptions.

This patch changes the initial exception vectors to use the count
register instead of the link register when we need to branch up to the
relocated kernel.

I have a dodgy userspace test which loops calling a function that reads
the PVR (mfpvr in userspace will be emulated by the kernel via the
program check exception).  On POWER8 and with CONFIG_RELOCATABLE=y, I
get a ~10% performance improvement with my userspace test with this
patch.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 7.0

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6192999

Upstream:
---------
https://git.kernel.org/cgit/linux/kernel/git/benh/powerpc.git/commit/?h=next&id=bc2e6c6ac21183a6102a926f83186d9cac6713f8

Test Status:
------------
Madhavan Srinivasan helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>