[fs] gfs2: Check for glock already held in gfs2_getxattr
Message-id: <1377168290.2720.25.camel@menhir>
Patchwork-id: 65807
O-Subject: [RHEL7 PATCH bz #997604] GFS2: Check for glock already held in gfs2_getxattr
Bugzilla: 997604
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This is taken directly from upstream:

http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=7bd9ee58a4fe026514266757e812cb89c3c945eb

It fixes a regression which was added by patch:

http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=6d4ade986f9c8df31e68fd30643997f79cc5a5f8

in which a recursive lock was taken during getxattr when called from the
selinux code during lookup. This fixes bug #997604

Steve.

----------------------------------------------------------------------------
>From 6c0d451c121bbd7f5a28836a725878c067395a1d Mon Sep 17 00:00:00 2001
From: Steven Whitehouse <swhiteho@redhat.com>
Date: Fri, 16 Aug 2013 21:10:28 +0100
Subject: [PATCH] GFS2: Check for glock already held in gfs2_getxattr

Since the introduction of atomic_open, gfs2_getxattr can be
called with the glock already held, so we need to allow for
this.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Reported-by: David Teigland <teigland@redhat.com>
Tested-by: David Teigland <teigland@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>