[fs] gfs2: Take glock reference in examine_bucket()
Message-id: <1377169469.2720.30.camel@menhir>
Patchwork-id: 65809
O-Subject: [RHEL7 PATCH bz #999897] GFS2: Race in iteration of glocks for unfreeze/umount
Bugzilla: 999897
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This patch has been queued for upstream here:

https://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/?id=7286b31eaba6404fa92f68d04626da1f395b3916

It fixes a race between gfs2_glock_hold() while iterating the glocks for
(glock) freeze or umount and the glock end of life process. The race
window is fairly small, so it is something which is not easy to hit in
general.

This was reported by, and the fix tested by Dave Teigland. It fixes bug
#999897,

Steve.

------------------------------------------------------------------------
>From 7286b31eaba6404fa92f68d04626da1f395b3916 Mon Sep 17 00:00:00 2001
From: Steven Whitehouse <swhiteho@redhat.com>
Date: Tue, 20 Aug 2013 09:35:09 +0100
Subject: [PATCH] GFS2: Take glock reference in examine_bucket()

We need to check the glock ref counter in a race free way
in order to ensure that the gfs2_glock_hold() call will
succeed. The easiest way to do that is to simply take the
reference count early in the common code of examine_bucket,
skipping any glocks with zero ref count.

That means that the examiner functions all need to put their
reference on the glock once they've performed their function.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Reported-by: David Teigland <teigland@redhat.com>
Tested-by: David Teigland <teigland@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>