[redhat] kabi: call stock find-provides for non-kernel subpackages
Message-id: <1377726347-12361-1-git-send-email-jarod@redhat.com>
Patchwork-id: 65922
O-Subject: [RHEL7 PATCH] [redhat] kabi: call stock find-provides for non-kernel subpackages
Bugzilla: 1000933
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>

Bugzilla: 1000933

Since we merged the kabi-whitelists bits, we're again calling a custom
kernel-provided find-provides that only parses kabi symbols. This was fine
in RHEL6, but in RHEL7, we now package kernel-tools, which includes a
-libs sub-package. The Provides: for libcpupower weren't getting properly
created. This should fix that.

Building in Brew right now to confirm functionality.

https://brewweb.devel.redhat.com/taskinfo?taskID=6222071

Signed-off-by: Jarod Wilson <jarod@redhat.com>