[redhat] kabi: fix find-provides for libraries for real now
Message-id: <1377783647-19919-1-git-send-email-jarod@redhat.com>
Patchwork-id: 65927
O-Subject: [RHEL7 PATCH] [redhat] kabi: fix find-provides for libraries for real now
Bugzilla: 1000933
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1000933

Since we merged the kabi-whitelists bits, we're again calling a custom
kernel-provided find-provides that only parses kabi symbols. This was fine
in RHEL6, but in RHEL7, we now package kernel-tools, which includes a
-libs sub-package. The Provides: for libcpupower weren't getting properly
created. The initial fix wasn't fully tested, and was lacking one very
critical piece, which was to pass the file list along. This time around,
I've actually had a chance to test locally outside of the rpm context, as
well as letting a brew build get far enough along to verify that the fix
carried all the way through to the actual rpm Provides:.

Still-in-progress-on-other-arches brew build verification:

$ rpm -qp --provides /mnt/redhat/brewroot/work/tasks/5054/6225054/kernel-tools-libs-3.10.0-13.el7.bz1000933.x86_64.rpm
kernel-tools-libs = 3.10.0-13.el7.bz1000933
kernel-tools-libs(x86-64) = 3.10.0-13.el7.bz1000933
libcpupower.so.0()(64bit)

Brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=6225051

Signed-off-by: Jarod Wilson <jarod@redhat.com>