[scsi] mpt2sas: when Async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Message-id: <1378391440-9452-14-git-send-email-thenzl@redhat.com>
Patchwork-id: 66233
O-Subject: [RHEL7 e-stor PATCH 13/16] [SCSI] mpt2sas: when Async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Bugzilla: 736230
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When Async scanning mode is enabled and device scanning is in progress then
devices should not be removed. But in actuality, devices are removed but
their transport layer entries are not removed. This causes error to add
the same device to the transport layer after host reset or diagnostic
reset.

So, in this patch, modified the code in such a way that device is not removed
when Async scanning mode is enabled and device scanning is in progress.

Signed-off-by: Sreekanth Reddy <Sreekanth.Reddy@lsi.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
(cherry picked from commit 4b7c081ca9dc4f97621138579dd4c4b29a7225fd)

Signed-off-by: Jarod Wilson <jarod@redhat.com>