[scsi] mpt3sas: Added a driver module parameter max_msix_vectors
Message-id: <1378391809-14103-9-git-send-email-thenzl@redhat.com>
Patchwork-id: 66232
O-Subject: [RHEL7 e-stor PATCH 8/9] [SCSI] mpt3sas: Added a driver module parameter max_msix_vectors
Bugzilla: 889435
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Added a driver module parameter max_msix_vectors. Using this module parameter
the maximum number of MSI-X vectors could be set.

The number of MSI-X vectors used would be the minimum of MSI-X vectors
supported by the HBA, the number of CPU cores and the value set to
max_msix_vectors module parameter.

The default value of this module parameter is set to 8. The default value of
this parameter is set to 8 inorder to reduce the amount of memory required for
Reply Descriptor Post queue.  This is because with the higher MSI-X vectors,
some times kernel is not able to allocate the requested amount of memory and
crash is observed. To overcome this problem, the default value is set to 8.

Signed-off-by: Sreekanth Reddy <Sreekanth.Reddy@lsi.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
(cherry picked from commit 4bd4a142107306e72e59f005c404ae33fd46509b)

Signed-off-by: Jarod Wilson <jarod@redhat.com>