[redhat] infra: eliminate redundant tarball creation
Message-id: <1378826436-27007-1-git-send-email-jarod@redhat.com>
Patchwork-id: 66592
O-Subject: [RHEL7 PATCH v3] [redhat] infra: eliminate redundant tarball creation
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

The rh-release and rh-dist-git targets were both triggering creation of a
kernel source tarball, as the commit hash of head was changing in between
the two, because rh-release wasn't committing changes before creating a
tarball. This was necessary in the rhel6 era, as redhat/lastcommit was used
to store a commit hash that was relied upon to properly generate the
changelog. We started storing a tag value in rhel7, but even that isn't
actually necessary, because we can figure out the prior tag with some
git rev-list and describe magic, implemented here.

Additionally, we'll stop calling make rh-srpm for the maintainer's export
to dist-git, and instead, just call rh-configs and sources-rh directly,
to skip creating an unnecessary srpm and to prevent a second spec file
and changelog update with the same contents.

Note: this change really only affects the maintainers, these targets are
never used by anyone but the maintainers, when they go to commit patchest
to the kernel git tree and export to dist-git.

Signed-off-by: Jarod Wilson <jarod@redhat.com>