[redhat] infra: enhance makefile to allow for configurable remote
Message-id: <1378828458-1587-1-git-send-email-nhorman@redhat.com>
Patchwork-id: 66594
O-Subject: [RHEL7 PATCH] redhat: enhance makefile to allow for configurable remote
Bugzilla:
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Veaceslav Falico <vfalico@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

I was noticing the other day that the rhg-* targets don't work properly when you
have multiple remotes configured and the origin remote isn't one brew can use.
To correct this, I modified the Makefile so that the RHGITURL variable was taken
from the rhg.url config element.  It still backs off to origin if that doesn't
exist, but this allows us to specify the specific url you want to pass to brew
in when running rhg-*.

Signed-off-by: Neil Horman <nhorman@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>