[netdrv] sfc: Ensure MCDI buffers, but not lengths, are dword aligned
Message-id: <1378913680-11601-24-git-send-email-nikolay@redhat.com>
Patchwork-id: 66801
O-Subject: [RHEL7 PATCH BZ 1005248 23/93] sfc: Ensure MCDI buffers, but not lengths, are dword aligned
Bugzilla: 1005248
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Ben Hutchings <bhutchings@solarflare.com>

We currently require that MCDI request and response lengths are
multiples of 4 bytes, because we will copy dwords in and out of shared
memory and we want to be sure we won't read or write out of bounds.
But all we really need to know is that there is sufficient padding for
that.  Also, we should ensure that buffers are dword-aligned, as on
some architectures misaligned access will result in data corruption or
a crash.

Change the buffer type to array-of-efx_dword_t and remove the
requirement that the lengths are multiples of 4.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
(cherry picked from commit 9528b9219348e0a013f4b587958a8ba9c96d7e20)

Signed-off-by: Jarod Wilson <jarod@redhat.com>