[redhat] Kconfig: enable CONFIG_NET_TULIP
Message-id: <1379008712-6661-1-git-send-email-gospo@redhat.com>
Patchwork-id: 67086
O-Subject: [PATCH RHEL7] Kconfig: enable CONFIG_NET_TULIP
Bugzilla: 1000953
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jesper Brouer <jbrouer@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

It was pointed out that some of the drivers I disabled earlier might be
needed as some hypervisors emulated hardware they supported.

QE noted that Hyper-V emulates tulip, so I'm going to enable it as Hyper-V is
one of the hypervisors we want to support.

This will resolve RHBZ 1000953.

Signed-off-by: Jarod Wilson <jarod@redhat.com>