[fs] cifs: don't instantiate new dentries in readdir for inodes that need to be revalidated immediately
Message-id: <1379097589-16967-6-git-send-email-sprabhu@redhat.com>
Patchwork-id: 67233
O-Subject: [RHEL7 PATCH 05/28] [bz1007981] cifs: don't instantiate new dentries in readdir for inodes that need to be revalidated immediately
Bugzilla: 1007981
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

David reported that commit c2b93e06 (cifs: only set ops for inodes in
I_NEW state) caused a regression with mfsymlinks. Prior to that patch,
if a mfsymlink dentry was instantiated at readdir time, the inode would
get a new set of ops when it was revalidated. After that patch, this
did not occur.

This patch addresses this by simply skipping instantiating dentries in
the readdir codepath when we know that they will need to be immediately
revalidated. The next attempt to use that dentry will cause a new lookup
to occur (which is basically what we want to happen anyway).

Cc: <stable@vger.kernel.org>
Cc: "Stefan (metze) Metzmacher" <metze@samba.org>
Cc: Sachin Prabhu <sprabhu@redhat.com>
Reported-and-Tested-by: David McBride <dwm37@cam.ac.uk>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 757c4f6260febff982276818bb946df89c1105aa)
Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>