[fs] xfs: Introduce ordered log vector support
Message-id: <1379494347-28456-20-git-send-email-dchinner@redhat.com>
Patchwork-id: 67341
O-Subject: [RHEL 7.0 PATCH 019/155] xfs: Introduce ordered log vector support
Bugzilla: 1001861
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RH BZ: 1001861
Upstream commit fd63875cc4cd60b9e5c609c24d75eaaad3e6d1c4
--

From: Dave Chinner <dchinner@redhat.com>

And "ordered log vector" is a log vector that is used for
tracking a log item through the CIL and into the AIL as part of the
log checkpointing. These ordered log vectors are special in that
they are not written to to journal in any way, and are not accounted
to the checkpoint being written.

The reason for this behaviour is to allow operations to attach items
to transactions and have them follow the normal transactional
lifecycle without actually having to write them to the journal. This
allows logging of items that track high level logical changes and
writing them to the log, while the physical items being modified
pass through into the AIL and pin the tail of the log (and therefore
the logical item in the log) until all the modified items are
physically written to disk.

IOWs, it allows us to write metadata without physically logging
every individual change but still maintain the full transactional
integrity guarantees we currently have w.r.t. crash recovery.

This change modifies some of the CIL item insertion loops, as
ordered log vectors introduce some new constraints as they don't
track any data. One advantage of this change is that it combines
two log vector chain walks into a single pass, so there is less
overhead in the transaction commit pass as well. It also kills some
unused code in the log vector walk loop when committing the CIL.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Mark Tinguely <tinguely@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>