[fs] xfs: Initialize all quota inodes to be NULLFSINO
Message-id: <1379494347-28456-43-git-send-email-dchinner@redhat.com>
Patchwork-id: 67352
O-Subject: [RHEL 7.0 PATCH 042/155] xfs: Initialize all quota inodes to be NULLFSINO
Bugzilla: 1001861
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chandra Seetharaman <sekharan@us.ibm.com>

RH BZ: 1001861
Upstream commit 0102629776a2f5c3bd9c5dc0ef77956c60255074
--

From: Chandra Seetharaman <sekharan@us.ibm.com>

mkfs doesn't initialize the quota inodes to NULLFSINO as it does for the
other internal inodes. This leads to two in-core values (0 and NULLFSINO)
to be checked against, to make sure if a quota inode is valid.

Solve that problem by initializing the in-core values of all quotaino
values to NULLFSINO if they are 0 in the disk.

Note that these values are not written back to on-disk superblock unless
some quota is enabled on the filesystem. Even in that case sb_pquotino is
written to disk only if the on-disk superblock supports pquotino

Signed-off-by: Chandra Seetharaman <sekharan@us.ibm.com>
Reviewed-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>